+++ testapps/GH931/src/main/java/gh931/AppEntry.java
@@ -1,0 +1,13 @@
+package gh931;
+
+import act.Act;
+import act.cli.Command;
+import act.cli.Optional;
+import act.inject.DefaultValue;
+
+public class AppEntry {
+
+    public static void main(String[] args) throws Exception {
+        Act.start();
+    }
+}